Add Maximo item 999 or Mas

+++ Kint-Sources-master/OpenNos.DAL.EF/BazaarItem.cs
@@ -20,7 +20,7 @@ namespace OpenNos.DAL.EF
     {
         #region Properties
 
-        public byte Amount { get; set; }
+        public ushort Amount { get; set; }
 
         public long BazaarItemId { get; set; }
